Reuse `assert_handle` in `call_function` and `evaluate`
